moved counter to its own component, added multiple counters with different titles passed as props
